refactor: Update configs
Update configs and table names in db
