--- auth/src/main/java/edu/pjwstk/auth/dto/response/AccessTokenResponse.java
@@ -1,6 +1,0 @@
-package edu.pjwstk.auth.dto.response;
-
-public record AccessTokenResponse(
-        String accessToken
-) {
-}
