+++ auth/src/main/java/edu/pjwstk/auth/util/OAuth2CodeUtil.java
@@ -1,10 +1,10 @@
-package edu.pjwstk.auth.services;
+package edu.pjwstk.auth.util;
 
 import edu.pjwstk.auth.dto.service.GoogleUserDto;
 
 import java.util.Map;
 
-public interface OAuth2CodeService {
+public interface OAuth2CodeUtil {
 
     Map<String, String> exchangeCodeForTokens(String code, String codeVerifier);
