+++ auth/src/main/java/edu/pjwstk/auth/util/impl/JwtTokenProvider.java
@@ -1,0 +1,93 @@
+package edu.pjwstk.auth.util.impl;
+
+import edu.pjwstk.auth.util.TokenProvider;
+import io.jsonwebtoken.Jwts;
+import org.apache.commons.codec.digest.DigestUtils;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+import javax.crypto.SecretKey;
+import javax.crypto.spec.SecretKeySpec;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+import java.util.UUID;
+
+public class JwtTokenProvider implements TokenProvider {
+
+    private final SecretKey secretKey;
+    private final UserDetailsService userDetailsService;
+    private final long accessTokenExpirationTime;
+    private final long refreshTokenExpirationTime;
+
+    public JwtTokenProvider(
+            UserDetailsService userDetailsService,
+            String secretKey,
+            long accessTokenExpirationTime,
+            long refreshTokenExpirationTime
+    ) {
+        this.secretKey = new SecretKeySpec(
+                secretKey.getBytes(StandardCharsets.UTF_8),
+                "HmacSHA256"
+        );
+        this.accessTokenExpirationTime = accessTokenExpirationTime;
+        this.userDetailsService = userDetailsService;
+        this.refreshTokenExpirationTime = refreshTokenExpirationTime;
+    }
+
+    @Override
+    public String generateToken(UUID userId, String email) {
+        return Jwts.builder()
+                .subject(email)
+                .claim("userId", userId)
+                .issuedAt(new Date())
+                .expiration(new Date(System.currentTimeMillis() + (accessTokenExpirationTime * 1000L)))
+                .signWith(secretKey)
+                .compact();
+    }
+
+    @Override
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parser()
+                    .verifyWith(secretKey)
+                    .build()
+                    .parseSignedClaims(token);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    @Override
+    public Authentication getAuthentication(String token) {
+        String email = Jwts.parser()
+                .verifyWith(secretKey)
+                .build()
+                .parseSignedClaims(token)
+                .getPayload()
+                .getSubject();
+
+        UserDetails userDetails = userDetailsService.loadUserByUsername(email);
+
+        return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+    }
+
+    @Override
+    public String generateRefreshToken() {
+        return UUID.randomUUID().toString();
+    }
+
+    @Override
+    public String hashToken(String token) {
+        return DigestUtils.sha256Hex(token);
+    }
+
+    @Override
+    public long getRefreshTokenExpirationTime() {
+        return refreshTokenExpirationTime;
+    }
+
+
+}
