+++ user/src/main/java/edu/pjwstk/user/persistence/UserEntity.java
@@ -18,7 +18,7 @@ import java.util.UUID;
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "users")
+@Table(name = "user")
 @DynamicUpdate // worth considering
 public class UserEntity {
     @Id
